fix: display groupchat avatar in header
