+++ src/services/type.ts
@@ -11,20 +11,7 @@ export interface MessagePart {
 export interface ChatResponse {
     messages: MessagePart[];
     reactionDelay: number; // 반응 지연시간 (ms)
-    characterState?: {
-        energy: number;     // 에너지 수준 (0~1)
-        mood: number;       // 기분 (0~1)
-        personality: {
-            agreeableness: number;      // 친화성
-            conscientiousness: number;  // 성실성
-            extroversion: number;       // 외향성
-            neuroticism: number;        // 신경성
-            openness: number;           // 개방성
-        };
-        socialBattery: number; // 사회적 배터리 (0~1)
-    };
-    // 모델이 구조화된 출력으로 제공하는 신규 메모리(선택)
-    newMemory?: string;
+    newMemory?: string; // 모델이 구조화된 출력으로 제공하는 신규 메모리(선택)
 }
 
 export interface GeminiApiPayload {
